+++ releases/dogwood/3/fun/config/lms/docker_run_production.py
@@ -140,6 +140,10 @@ STATIC_URL = "/static/"
 MEDIA_ROOT = path("/edx/var/edxapp/media/")
 MEDIA_URL = "/media/"
 
+STATICFILES_STORAGE = config(
+    "STATICFILES_STORAGE", default="lms.envs.fun.storage.LMSManifestStaticFilesStorage"
+)
+
 # DEFAULT_COURSE_ABOUT_IMAGE_URL specifies the default image to show for courses that don't provide one
 DEFAULT_COURSE_ABOUT_IMAGE_URL = config(
     "DEFAULT_COURSE_ABOUT_IMAGE_URL", default=DEFAULT_COURSE_ABOUT_IMAGE_URL